refactor: renames profile to impact
