+++ src/screens/users/ImpactScreen/CommunityDonationsImpactCards/ImpactDonationsVector.tsx
