+++ src/screens/users/ImpactScreen/DirectDonationsImpactCards/index.tsx
@@ -3,20 +3,18 @@ import { useCallback } from "react";
 import { useNavigation } from "hooks/useNavigation";
 import { useTranslation } from "react-i18next";
 import usePersonPayments from "hooks/apiHooks/usePersonPayments";
-import { formatDateTime } from "lib/formatters/dateFormatter";
-import CardImageText from "components/moleculars/CardImageText";
 import { theme } from "@ribon.io/shared/styles";
-import { formatPrice } from "lib/formatters/currencyFormatter";
+import DirectDonationCard from "screens/users/ImpactScreen/DirectDonationsImpactCards/DirectDonationCard";
 import { logEvent } from "services/analytics";
 import { useFocusEffect } from "@react-navigation/native";
 import ImpactDonationsVector from "./ImpactDonationsVector";
 import S from "./styles";
 import ZeroDonationsSection from "../ZeroDonationsSection";
 
-function CommunityDonationsImpactCards(): JSX.Element {
-  const { useCommunityPersonPayments } = usePersonPayments();
+function DirectDonationsImpactCards(): JSX.Element {
+  const { useDirectPersonPayments } = usePersonPayments();
 
-  const { data, refetch } = useCommunityPersonPayments(1, 6);
+  const { data, refetch } = useDirectPersonPayments(1, 6);
 
   useFocusEffect(
     useCallback(() => {
@@ -28,29 +26,19 @@ function CommunityDonationsImpactCards(): JSX.Element {
   const hasImpact = impactItems() && impactItems()?.length > 0;
   const { navigateTo } = useNavigation();
   const { t } = useTranslation("translation", {
-    keyPrefix: "users.profileScreen.ngoImpactCards.zeroDonationsSection",
+    keyPrefix: "users.impactScreen.ngoImpactCards.zeroDonationsSection",
   });
 
   const navigateToPromotersScreen = () => {
-    logEvent("giveCauseCard_click", { from: "impactEmptystate" });
-    navigateTo("PromotersScreen");
+    logEvent("giveNonProfitCard_click", { from: "impactEmptystate" });
+    navigateTo("PromotersScreen", { isInCommunity: false });
   };
 
   const impactCardsList = () => (
     <View style={S.cardsContainer}>
       {impactItems()?.map((item) => (
         <View key={item?.id} style={{ marginBottom: theme.spacingNative(12) }}>
-          <CardImageText
-            subtitle={item.receiver.name}
-            title={
-              item.offer
-                ? formatPrice(item.offer.priceValue, item.offer.currency)
-                : `${item.amountCents / 100} USDC`
-            }
-            footerText={formatDateTime(item.paidDate)}
-            subtitleStyle={S.subtitleStyle}
-            titleStyle={S.titleStyle}
-          />
+          <DirectDonationCard personPayment={item} />
         </View>
       ))}
     </View>
@@ -60,13 +48,13 @@ function CommunityDonationsImpactCards(): JSX.Element {
     impactCardsList()
   ) : (
     <ZeroDonationsSection
-      title={t("community.title")}
+      title={t("direct.title")}
       onButtonPress={navigateToPromotersScreen}
-      description={t("community.description")}
-      buttonText={t("community.buttonText")}
+      description={t("direct.description")}
+      buttonText={t("direct.buttonText")}
       image={<ImpactDonationsVector />}
     />
   );
 }
 
-export default CommunityDonationsImpactCards;
+export default DirectDonationsImpactCards;
