+++ src/screens/users/ImpactScreen/NgoImpactCards/styles.ts
