+++ src/types/index.tsx
@@ -47,7 +47,7 @@ export type PostDonationScreenParams = {
 export type RootTabParamList = {
   CausesScreen: undefined;
   ForYouScreen: undefined;
-  ProfileScreen: undefined;
+  ImpactScreen: undefined;
   PromotersScreen: PromotersScreenParams;
   PostDonationScreen: PostDonationScreenParams;
 };
